infrastructure/src/main/resources/prediction_container/requirements.txt
google-cloud-pubsub==2.5.0
pillow==9.0.0
tf-agents==0.8.0
tensorflow==2.5.3